factor out socks-headers.ts
